Refactor rest of utils
